fix: resolve 8 open issues via multi-agent parallel implementation
Web Avatar (#77, #76):
- Create shared entity-utils.js module for centralized avatar/name resolution
- Replace hardcoded ENTITY_CHARS (0-3 only) across 5 portal pages
- Entity IDs 4-7 now show proper avatars instead of '?'
- Chat/dashboard/files/mission/schedule all use server-synced avatars

Android Entity (#72, #73):
- Sync serverEntityLimit in MainActivity.loadBoundEntities()
- Extend badge colors for entity IDs 4-7 in ClawRenderer
- Fix Float precision loss: change parts Map<String,Float> to Map<String,Double>
- Update all ClawRenderer parts consumers with explicit .toFloat() calls

Web UX (#79, #80, #75):
- Feedback page: use standard .container/.card classes, add --bg-secondary CSS var
- Schedule page: add showEditDialog() for editing existing schedules
- Add lobster emoji favicon to all 13 portal HTML files

Bot API (#74):
- Add getMissionApiHints() helper with pre-filled curl templates
- Include [AVAILABLE TOOLS] section in all 6 push notification paths
- Bots now always see Mission Dashboard API hints in every message

+++ app/src/main/java/com/hank/clawlive/data/model/AgentStatus.kt
@@ -5,7 +5,7 @@ data class AgentStatus(
     val character: String = "LOBSTER",
     val state: CharacterState = CharacterState.IDLE,
     val message: String = "Loading...",
-    val parts: Map<String, Float>? = null,
+    val parts: Map<String, Double>? = null,
     val lastUpdated: Long = System.currentTimeMillis(),
     val isBound: Boolean = false,
     val usage: UsageInfo? = null,
