+++ app/src/main/java/com/hank/clawlive/data/model/EntityStatus.kt
@@ -10,7 +10,7 @@ data class EntityStatus(
     val character: String = "LOBSTER",
     val state: CharacterState = CharacterState.IDLE,
     val message: String = "Loading...",
-    val parts: Map<String, Float>? = null,
+    val parts: Map<String, Double>? = null,
     val lastUpdated: Long = System.currentTimeMillis(),
     val isBound: Boolean = false,
     val usage: UsageInfo? = null,
